+++ frontend/src/lib/components/apps/types.ts
@@ -176,7 +176,6 @@ export type ListContext = Writable<{
 export type ListInputs = {
 	set: (id: string, value: any) => void
 	remove: (id: string) => void
-	get: (id: string) => void
 }
 
 export type GroupContext = Writable<Record<string, any>>