+++ frontend/src/lib/components/runs/JobPreview.svelte
@@ -36,7 +36,7 @@
 
 <TestJobLoader bind:job={currentJob} bind:watchJob on:done={onDone} />
 
-<div class="p-4 flex flex-col gap-2 items-start">
+<div class="p-4 flex flex-col gap-2 items-start h-full">
 	{#if job}
 		<div class="flex gap-2">
 			{#if job?.['priority']}
@@ -77,7 +77,7 @@
 			</div>
 		{/if}
 
-		<div class=" w-full rounded-md overflow-auto">
+		<div class=" w-full rounded-md min-h-full">
 			{#if job?.type === Job.type.COMPLETED_JOB}
 				<Tabs bind:selected={viewTab}>
 					<Tab size="xs" value="result">Result</Tab>
@@ -96,7 +96,7 @@
 							</div>
 						</div>
 					{:else}
-						<div class="flex flex-row border rounded-md p-2 mt-2 max-h-1/2 overflow-auto">
+						<div class="flex flex-col border rounded-md p-2 mt-2 h-full overflow-auto">
 							{#if viewTab == 'logs'}
 								<div class="w-full">
 									<LogViewer