+++ frontend/src/routes/(root)/(logged)/workspace_settings/+page.svelte
@@ -58,7 +58,10 @@
 	let errorHandlerExtraArgs: Record<string, any> = {}
 	let errorHandlerMutedOnCancel: boolean | undefined = undefined
 	let openaiResourceInitialPath: string | undefined = undefined
-	let s3ResourceInitialPath: string | undefined = undefined
+	let s3ResourceSettings: {
+		s3ResourcePath: string | undefined
+		publicResource: boolean | undefined
+	}
 	let gitSyncSettings: {
 		script_path: string
 		git_repo_resource_path: string
@@ -171,16 +174,16 @@
 		sendUserToast(`Copilot settings updated`)
 	}
 
-	async function editWindmillLFSSettings(s3ResourcePath: string): Promise<void> {
-		s3ResourceInitialPath = s3ResourcePath
-		if (s3ResourcePath) {
-			let resourcePathWithPrefix = `$res:${s3ResourcePath}`
+	async function editWindmillLFSSettings(): Promise<void> {
+		if (!emptyString(s3ResourceSettings.s3ResourcePath)) {
+			let resourcePathWithPrefix = `$res:${s3ResourceSettings.s3ResourcePath}`
 			await WorkspaceService.editLargeFileStorageConfig({
 				workspace: $workspaceStore!,
 				requestBody: {
 					large_file_storage: {
 						type: LargeFileStorage.type.S3STORAGE,
-						s3_resource_path: resourcePathWithPrefix
+						s3_resource_path: resourcePathWithPrefix,
+						public_resource: s3ResourceSettings.publicResource
 					}
 				}
 			})
@@ -280,10 +283,17 @@
 		errorHandlerExtraArgs = settings.error_handler_extra_args ?? {}
 		codeCompletionEnabled = settings.code_completion_enabled
 		workspaceDefaultAppPath = settings.default_app
-		s3ResourceInitialPath =
+
+		s3ResourceSettings =
 			settings.large_file_storage?.type === LargeFileStorage.type.S3STORAGE
-				? settings.large_file_storage?.s3_resource_path?.replace('$res:', '')
-				: undefined
+				? {
+						s3ResourcePath: settings.large_file_storage?.s3_resource_path?.replace('$res:', ''),
+						publicResource: settings.large_file_storage?.public_resource
+				  }
+				: {
+						s3ResourcePath: undefined,
+						publicResource: undefined
+				  }
 		if (
 			settings.git_sync !== undefined &&
 			settings.git_sync !== null &&
@@ -762,25 +772,48 @@
 				</Alert>
 			{/if}
 			<div class="mt-5 flex gap-1">
-				{#key s3ResourceInitialPath}
-					<ResourcePicker
-						resourceType="s3"
-						initialValue={s3ResourceInitialPath}
-						on:change={(ev) => {
-							editWindmillLFSSettings(ev.detail)
-						}}
-					/>
+				{#key s3ResourceSettings.s3ResourcePath}
+					<ResourcePicker resourceType="s3" bind:value={s3ResourceSettings.s3ResourcePath} />
 				{/key}
 				<Button
 					size="sm"
 					variant="contained"
 					color="dark"
-					disabled={!s3ResourceInitialPath}
+					disabled={emptyString(s3ResourceSettings.s3ResourcePath)}
 					on:click={async () => {
 						if ($workspaceStore) {
 							s3FileViewer?.open?.(undefined)
 						}
 					}}>Browse content</Button
+				>
+			</div>
+			<div class="flex flex-col mt-5 mb-1 gap-1">
+				<Toggle
+					disabled={emptyString(s3ResourceSettings.s3ResourcePath)}
+					bind:checked={s3ResourceSettings.publicResource}
+					options={{
+						right: 'S3 resource details can be accessed by all users of this workspace',
+						rightTooltip:
+							'If set, all users of this workspace will have access the to entire content of the S3 bucket, as well as the resource details. this effectively by-pass the permissions set on the resource and makes it public to everyone.'
+					}}
+				/>
+				{#if s3ResourceSettings.publicResource === true}
+					<Alert type="warning" title="S3 bucket content and resource details are shared">
+						S3 resource public access is ON, which means that the entire content of the S3 bucket
+						will be accessible to all the users of this workspace regardless of whether they have
+						access the resource or not. Similarly, certain Windmill SDK endpoints can be used in
+						scripts to access the resource details, including public and private keys.
+					</Alert>
+				{/if}
+			</div>
+			<div class="flex mt-5 mb-5 gap-1">
+				<Button
+					color="blue"
+					disabled={emptyString(s3ResourceSettings.s3ResourcePath)}
+					on:click={() => {
+						editWindmillLFSSettings()
+						console.log('Saving S3 settings', s3ResourceSettings)
+					}}>Save S3 settings</Button
 				>
 			</div>
 		{:else if tab == 'git_sync'}
@@ -813,7 +846,18 @@
 				<br />
 				Filtering out certain sensitive folders from the sync will be available soon.
 			</Alert>
-
+			{#if $enterpriseLicense}
+				<div class="flex mt-5 mb-5 gap-1">
+					<Button
+						color="blue"
+						disabled={gitSyncSettings?.some((elmt) => emptyString(elmt.git_repo_resource_path))}
+						on:click={() => {
+							editWindmillGitSyncSettings()
+							console.log('Saving git sync settings', gitSyncSettings)
+						}}>Save Git sync settings</Button
+					>
+				</div>
+			{/if}
 			{#if Array.isArray(gitSyncSettings)}
 				{#each gitSyncSettings as gitSyncSettingsElmt, idx}
 					<div class="flex mt-5 mb-1 gap-1 items-center text-xs">
@@ -893,7 +937,7 @@
 						gitSyncSettings = [
 							...gitSyncSettings,
 							{
-								script_path: 'hub/7927/sync-script-to-git-repo-windmill',
+								script_path: 'hub/7929/sync-script-to-git-repo-windmill',
 								git_repo_resource_path: '',
 								use_individual_branch: false
 							}
@@ -906,6 +950,20 @@
 							}
 						]
 					}}>Add connection</Button
+				>
+				<Button
+					color="none"
+					variant="border"
+					on:click={() => {
+						gitSyncSettings = [...gitSyncSettings.slice(0, -1)]
+						gitSyncTestJobs = [
+							...gitSyncTestJobs,
+							{
+								jobId: undefined,
+								status: undefined
+							}
+						]
+					}}>Delete connection</Button
 				>
 			</div>
 
@@ -944,16 +1002,6 @@ git push</code
 						></pre
 					>
 				</div>
-			</div>
-			<div class="flex mt-5 mb-5 gap-1">
-				<Button
-					color="blue"
-					disabled={gitSyncSettings?.some((elmt) => emptyString(elmt.git_repo_resource_path))}
-					on:click={() => {
-						editWindmillGitSyncSettings()
-						console.log('Saving git sync settings', gitSyncSettings)
-					}}>Save Git sync settings</Button
-				>
 			</div>
 		{:else if tab == 'default_app'}
 			<PageHeader